fixed up tool calls a lil
made it so when it runs the tool call, you don't see the thought process and the actual call syntax of the tool. its good for debugging but i got it working so i removed the log for it. also changed up the name to fit as a shell command