added voice activation and other stuff that i am going to mention in description
this took longer than expected, because i had to redo a few things and figure out how i want things to work. i made it so to activate the llm you say jarvis activate so you can talk normally. even after u activate the llm you can still talk normally because to talk to it you say jarvis, (the thing you want to say). i did this because if i want to add hard coded non llm jarvis commands i should make it so the llm isnt always on. i also added a requirements.txt for people without the speech recognition library (and requests) and then updated the readme with additional instructions and clarifications. i also added a micindexes.py file for verifiying the mic index to use for the voice activation. it lists all the input devices connected and their index number. find the index number that correlates to the microphone you are using and input that when you run main.py. this took a while hope it works well. will prolly do a few more things today but might just go straight to bed.