fixed issues and added tts
it would sometimes do unknownvalueerror if u weren't explicitly clear in your text, so i just made it continue and since it loops it will just get ur next prompt. i also added tts in form of the pyttsx3 library for python. it will read out the actual response of the ai after you ask it. it has also been added to the requirements.txt. it will not read out the think tag stuff because those are the thoughts of the ai.