added memory saving tools and a lot mroe
added ai remember bool for the system prompt and what im about to talk about not really sure if the system prompt part even works but if it aint broke dont fix it. i added 2 new tool calls one that turns off memory saving and this turns off the airemember bool which inside of main makes it so it doesnt append the memory history if airemember is false. i also made it so if u run this tool it resets the memory to avoid any issues. i did the exact opposite for turnonmemorysaving tool. it turns on the bool, but still clears the memory because i was running into some issues. i also changed the way it detects tool calls. the syntax is still the same for the toolcall but the way i find it is different because i dont like the way 're' works. ok thats it bye