Moved around and optimized the code, also added new tool and other stuff to make tool work.
made a new function called main and renamed the old main. i made a bool to check if it is running (the llm phase) so that it isnt cooked (i might be able to just remove the bool and use while True but ion know) i also made use of the phase string and now when u first run the file, it runs main and it defaults to non llm but when you call main again and right before u switch the phase, it goes into another phase and i made that into a tool call. when you say 'jarvis turn off' it goes into non llm mode which wont do anything until you say jarvis activate and it will reset the ais previous memory and bring you into jarvis llm mode. i also optimized the code a little bit and moved things around.